REFACTOR: organizing files
